feat(ingest): s3 - add support for multiple pathspecs in one recipe (#4777)
