+++ metadata-ingestion/archived/source_docs/s3.md
@@ -1,8 +1,8 @@
-# S3
+# S3 Crawling with Glue
 
 :::note
 
-Our [data-lake connector](./data_lake.md) allows you to ingest S3 files directly with the option to compute profiles as well. The following guide describes how to ingest S3 datasets through cataloging them in AWS Glue.
+Our [S3 data lake](./s3_data_lake.md) source allows you to ingest S3 files directly with the option to compute profiles as well. The following guide describes how to ingest S3 datasets through cataloging them in AWS Glue.
 
 :::
