+++ metadata-ingestion/docs/sources/s3/s3.md
@@ -1,4 +1,4 @@
-## Valid path_spec.include
+## Valid path_specs.include
 
 ```python
 s3://my-bucket/foo/tests/bar.avro # single file table   
@@ -6,7 +6,7 @@ s3://my-bucket/foo/tests/*.* # mulitple file level tables
 s3://my-bucket/foo/tests/{table}/*.avro #table without partition
 s3://my-bucket/foo/tests/{table}/*/*.avro #table where partitions are not specified
 s3://my-bucket/foo/tests/{table}/*.* # table where no partitions as well as data type specified
-s3://my-bucket/{dept}/tests/{table}/*.avro # specifying key wards to be used in display name
+s3://my-bucket/{dept}/tests/{table}/*.avro # specifying keywords to be used in display name
 s3://my-bucket/{dept}/tests/{table}/{partition_key[0]}={partition[0]}/{partition_key[1]}={partition[1]}/*.avro # specify partition key and value format
 s3://my-bucket/{dept}/tests/{table}/{partition[0]}/{partition[1]}/{partition[2]}/*.avro # specify partition value only format
 s3://my-bucket/{dept}/tests/{table}/{partition[0]}/{partition[1]}/{partition[2]}/*.* # for all extensions
@@ -14,7 +14,7 @@ s3://my-bucket/*/{table}/{partition[0]}/{partition[1]}/{partition[2]}/*.* # tabl
 s3://my-bucket/*/*/{table}/{partition[0]}/{partition[1]}/{partition[2]}/*.* # table is present at 3 levels down in bucket
 ```
 
-## Valid path_spec.exclude
+## Valid path_specs.exclude
 - **/tests/**
 - s3://my-bucket/hr/**
 - **/tests/*.csv
@@ -40,7 +40,7 @@ If you would like to write a more complicated function for resolving file names,
 
 :::caution
 
-Specify as long fixed prefix ( with out /*/ ) as possible in `path_spec.include`. This will reduce the scanning time and cost, specifically on AWS S3 
+Specify as long fixed prefix ( with out /*/ ) as possible in `path_specs.include`. This will reduce the scanning time and cost, specifically on AWS S3
 
 :::
