+++ metadata-ingestion/tests/integration/s3/golden-files/local/golden_mces_folder_no_partition_glob.json
@@ -7,7 +7,10 @@
             "aspects": [
                 {
                     "com.linkedin.pegasus2avro.dataset.DatasetProperties": {
-                        "customProperties": {},
+                        "customProperties": {
+                            "number_of_files": "2",
+                            "size_in_bytes": "8412"
+                        },
                         "externalUrl": null,
                         "name": "food_parquet",
                         "qualifiedName": null,
