+++ metadata-ingestion/tests/integration/s3/golden-files/local/golden_mces_folder_partition_basic.json
@@ -7,7 +7,10 @@
             "aspects": [
                 {
                     "com.linkedin.pegasus2avro.dataset.DatasetProperties": {
-                        "customProperties": {},
+                        "customProperties": {
+                            "number_of_files": "12",
+                            "size_in_bytes": "174060"
+                        },
                         "externalUrl": null,
                         "name": "folder_aaa.pokemon_abilities_json",
                         "qualifiedName": null,
