+++ metadata-ingestion/tests/integration/s3/sources/s3/file_without_extension.json
@@ -1,11 +1,11 @@
 {
   "type": "s3",
   "config": {
-    "env": "UAT",
+    "platform_instance": "test-platform-instance",
+    "env": "DEV",
     "path_specs": [{
-      "include": "s3://my-test-bucket/folder_a/folder_aa/{dept}/{table}/*.*",
-      "file_types": ["csv"],
-      "table_name": "{dept}.{table}"
+      "include": "s3://my-test-bucket/folder_a/folder_aa/folder_aaa/no_extension/*",
+      "default_extension": "csv"
     }],
     "aws_config": {
       "aws_region": "us-east-1",
