+++ metadata-ingestion/tests/integration/s3/sources/s3/folder_no_partition.json
@@ -2,9 +2,9 @@
   "type": "s3",
   "config": {
     "env": "UAT",
-    "path_spec": {
+    "path_specs": [{
       "include": "s3://my-test-bucket/folder_a/folder_aa/folder_aaa/{table}/*.*"
-    },
+    }],
     "aws_config": {
       "aws_region": "us-east-1",
       "aws_access_key_id": "testing",
