+++ metadata-ingestion/tests/integration/s3/sources/s3/folder_partition_keyval.json
@@ -2,10 +2,10 @@
   "type": "s3",
   "config": {
     "env": "UAT",
-    "path_spec": {
+    "path_specs": [{
       "include": "s3://my-test-bucket/folder_a/folder_aa/{dept}/{table}/{partition_key[0]}={partition[0]}/{partition_key[1]}={partition[1]}/*.*",
       "table_name": "{dept}.{table}"
-    },
+    }],
     "aws_config": {
       "aws_region": "us-east-1",
       "aws_access_key_id": "testing",
