+++ metadata-ingestion/tests/integration/s3/sources/s3/multiple_files.json
@@ -4,9 +4,9 @@
     "platform": "test-platform",
     "platform_instance": "test-platform-instance",
     "env": "DEV",
-    "path_spec": {
+    "path_specs": [{
       "include": "s3://my-test-bucket/folder_a/folder_aa/folder_aaa/*.*"
-    },
+    }],
     "aws_config": {
       "aws_region": "us-east-1",
       "aws_access_key_id": "testing",
