+++ metadata-ingestion/tests/integration/s3/test_data/local_system/folder_a/folder_aa/folder_aaa/no_extension/small
@@ -1,0 +1,6 @@
+"1st chord","2nd chord","3rd chord","4th chord","Progression Quality"
+6,4,1,5,"Alternative",
+4,4,1,5,"Catchy",
+1,1,1,1,"Didgeridoo",
+1,6,4,5,"Dreadful",
+1,6,2,5,"Dreadful",
